pyproject.toml
[project]
name = "simple-a2a-agent"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "a2a-sdk[all]>=0.3.22",
    "pydantic-ai>=1.58.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0",
  "pytest-cov>=4.0",
  "mypy>=1.10",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

# Ruff configuration
[tool.ruff]
line-length = 119
target-version = "py312"

[tool.ruff.lint]
select = [
  "E4", "E7", "E9",  # Errors
  "F",               # PyFlakes
  "B",               # Bugbear
  "W",               # Warnings
  "I",               # isort
  "C4",              # Comprehensions
  "UP",              # Pyupgrade
]
ignore = [
  "E501",  # Line length (handled by formatter)
]

[tool.ruff.lint.isort]
section-order = [
  "future",
  "standard-library",
  "third-party",
  "first-party",
  "local-folder",
]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

# MyPy configuration
[tool.mypy]
python_version = "3.12"
strict = false
warn_return_any = true
warn_unused_configs = true
exclude = [
  "migrations/",
  "tests/",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]

[dependency-groups]
dev = [
    "pytest>=8.0",
    "pytest-cov>=4.0",
    "ruff>=0.15.0",
    "ty>=0.0.16",
]
